start a `jupyter-serde` crate with the media portion of runtimelib
