+++ Cargo.toml
@@ -3,6 +3,7 @@
 members = [
     "crates/nbformat",
     "crates/runtimelib",
+    "crates/jupyter-serde",
     "crates/jupyter-websocket-client",
 ]
