+++ crates/jupyter-serde/Cargo.toml
@@ -1,15 +1,14 @@
 [package]
-name = "nbformat"
+name = "jupyter-serde"
 version = "0.1.1"
 edition = "2021"
-description = "Parse Jupyter Notebooks"
+description = "Serialization and Deserialization for Jupyter related data structures"
 repository = "https://github.com/runtimed/runtimed"
 license = "BSD-3-Clause"
 
 [dependencies]
 chrono = "0.4"
 anyhow = "1.0.80"
-jupyter-serde = { version = "0.1.0", path = "../jupyter-serde" }
 serde_json = "1.0.113"
 serde = { version = "1.0.196", features = ["derive"] }
 thiserror = "1.0"
