+++ crates/jupyter-serde/src/media/datatable.rs
