+++ crates/runtimelib/Cargo.toml
@@ -21,6 +21,7 @@ data-encoding = "2.5.0"
 dirs = "5.0.1"
 smol = { version = "1.2", optional = true }
 futures = { version = "0.3" }
+jupyter-serde = { version = "0.1.0", path = "../jupyter-serde" }
 rand = "0.8.5"
 ring = "0.17.7"
 serde = { version = "1.0.196", features = ["derive"] }
