+++ crates/runtimelib/src/lib.rs
@@ -1,7 +1,8 @@
-pub mod jupyter;
-pub mod media;
-pub mod messaging;
+pub use jupyter_serde::media;
+pub use jupyter_serde::media::*;
+pub use jupyter_serde::ExecutionCount;
 
+pub mod jupyter;
+pub mod messaging;
 pub use jupyter::*;
-pub use media::*;
 pub use messaging::*;
